added fluid synth
git-svn-id: https://svn.code.sf.net/p/pianobooster/code/trunk@66 ba081f5d-443b-49a7-ac4b-446c3f91f371

+++ PianoBooster/src/Scroll.cpp
@@ -74,8 +74,8 @@ bool CScroll::insertSlots()
 
     while (true)
     {
-        int headDelta = deltaAdjust(m_deltaHead) * m_noteSpacingFactor;
-        int slotDetlta = Cfg::staveEndX() - Cfg::playZoneX() - m_headSlot.getDeltaTime() * m_noteSpacingFactor - NOTE_BEHIND_GAP;
+        float headDelta = deltaAdjust(m_deltaHead) * m_noteSpacingFactor;
+        float slotDetlta = Cfg::staveEndX() - Cfg::playZoneX() - m_headSlot.getDeltaTime() * m_noteSpacingFactor - NOTE_BEHIND_GAP;
 
         if (headDelta > slotDetlta)
             break;
